move back GzeFollowerIndicator into colors package
