+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/ColorToolBox.java
@@ -27,7 +27,6 @@ import net.gazeplay.commons.ui.Translator;
 import net.gazeplay.commons.utils.games.GazePlayDirectories;
 import net.gazeplay.components.AbstractGazeIndicator;
 import net.gazeplay.components.CssUtil;
-import net.gazeplay.components.GazeFollowerIndicator;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
