+++ gazeplay-games/src/main/java/net/gazeplay/games/colors/CustomColorPicker.java
@@ -18,7 +18,6 @@ import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.IGameContext;
 import net.gazeplay.components.AbstractGazeIndicator;
-import net.gazeplay.components.GazeFollowerIndicator;
 
 import static net.gazeplay.games.colors.ColorToolBox.COLORIZE_BUTTONS_SIZE_PX;
 import static net.gazeplay.games.colors.ColorToolBox.COLORS_IMAGES_PATH;
